+++ bubble/list/list.go
@@ -81,8 +81,8 @@ type Model struct {
 
 func New(delegate ItemDelegate, config *settings.Config, cat *categories.Categories, favorites *favorites.Favorites, width, height int) *Model {
 	return newModel(delegate, config, cat, favorites, width, height,
-		getService(config.DebugMode),
-		getHistory(config.DebugMode, config.DoNotMarkSubmissionsAsRead))
+		getService(config.DebugMode, config.DebugFallible),
+		getHistory(config.DebugMode || config.DebugFallible, config.DoNotMarkSubmissionsAsRead))
 }
 
 func newModel(delegate ItemDelegate, config *settings.Config, cat *categories.Categories, favorites *favorites.Favorites, width, height int, service hn.Service, hist history.History) *Model {